What used to be float literals are now double literals. Float literals have the "f" suffix.

+++ spec/crystal/spec/parser/parser_spec.cr
@@ -25,7 +25,7 @@ class Int
   end
 end
 
-class Float
+class Double
   def float
     FloatLiteral.new to_s
   end
